fix add team integration

+++ database/migrations/2024_09_18_150652_add_team_id_to_categories_table.php
@@ -11,10 +11,13 @@ return new class extends Migration
      */
     public function up(): void
     {
-        Schema::table('categories', function (Blueprint $table) {
-            $table->unsignedBigInteger('team_id')->nullable()->after('id');
-            $table->foreign('team_id')->references('id')->on('teams')->onDelete('cascade');
-        });
+        if(Schema::hasTable('teams')){
+            Schema::table('categories', function (Blueprint $table) {
+                $table->unsignedBigInteger('team_id')->nullable()->after('id');
+                $table->foreign('team_id')->references('id')->on('teams')->onDelete('cascade');
+            });
+        }
+
     }
 
     /**
@@ -22,9 +25,11 @@ return new class extends Migration
      */
     public function down(): void
     {
-        Schema::table('categories', function (Blueprint $table) {
-            $table->dropForeign(['team_id']);
-            $table->dropColumn('team_id');
-        });
+        if(Schema::hasTable('teams')) {
+            Schema::table('categories', function (Blueprint $table) {
+                $table->dropForeign(['team_id']);
+                $table->dropColumn('team_id');
+            });
+        }
     }
 };
